Refactor reading pasive items
