+++ SpaceManager/view/Components/ReadItem/ReadItem.swift
@@ -1,0 +1,18 @@
+//
+//  ReadItem.swift
+//  SpaceManager
+//
+//  Created by Kuba Kromomołowski on 29/08/2024.
+//
+
+import SwiftUI
+
+struct ReadItem: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    ReadItem()
+}
